requirements.txt
numpy>=1.20.0
scikit-learn>=1.0.0
requests>=2.25.0
typer>=0.9.0
sympy>=1.12
python-dotenv>=1.0.0
